rootReducer component created under reducers => index.js. Getting redux enviornment ready to implement logic and finish final boilerplating
