+++ src/Redux/actions/types.js
@@ -1,4 +1,10 @@
-// LOGIN ACTION constant types 
+// LOGIN ACTION constant types
 export const LOGIN_START = 'LOGIN_START';
 export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
 export const LOGIN_ERROR = 'LOGIN_ERROR';
+
+// TICKET ACTIONS constant types
+export const TICKET_NEW = 'TICKET_NEW';
+export const TICKET_UPDATE = 'TICKET_UPDATE';
+export const TICKET_ERROR = 'TICKET_ERROR';
+export const TICKET_ADD = 'TICKET_ADD';
