+++ src/Redux/reducers/index.js
@@ -1,2 +1,9 @@
-// combining reducers
-// root reducer home
+import { combineReducers } from 'redux';
+import loginReducer from './loginReducer';
+// import ticketReducer from './ticketReducer';
+
+// calling the default reducer to create a link.
+
+const rootReducer = combineReducers({ loginReducer });
+
+export default rootReducer;
