+++ src/Redux/store.js
@@ -1,13 +1,8 @@
-import { createStore, applyMiddleware, combineReducers } from 'redux';
+import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
+import rootReducer from './reducers';
 
-import loginReducer from './reducers/loginReducer';
-// import ticketReducer from './redux/reducers/ticketReducer';
-
-// calling the default reducer to create a link.
-const rootReducer = combineReducers({ loginReducer });
-// ticketReducer
 const store = createStore(rootReducer, applyMiddleware(thunk, logger));
 
 export default store;
